+++ app/src/app/components/products/form.component.html
@@ -18,12 +18,10 @@
         </div>
         <div class="form-group row">
           <div class="col-sm-6">
-            <label for="price">Precio Unico</label>
             <input type="number" id="price" class="form-control form-control-rounded"
             pattern='[0-9]+(\.[0-9][0-9]?)?' formControlName="unitPrice" placeholder="Precio Unitario" name="priceUnit">
           </div>
           <div class="col-sm-6">
-            <label for="quantity">Cantidad</label>
             <input type="number" id="quantity" class="form-control form-control-rounded" placeholder="Cantidad"
             formControlName="quantity" name="quantity">
           </div>
